fashion_trend_app/requirements.txt
# Fashion Trend Analysis App - Requirements

# Core Flask dependencies
Flask>=2.3.0
Werkzeug>=2.3.0
Jinja2>=3.1.0

# Machine Learning
tensorflow>=2.16.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Data handling
scipy>=1.11.0

# Development
python-dotenv>=1.0.0
